requirements.txt
boto==2.33.0
gnureadline==6.3.3
ipython==2.3.0
python-dateutil==2.2
six==1.8.0
wsgiref==0.1.2
